refactor: decompose rim_slope_match_out.py and optimize diagnostic hot-loops
