+++ modules/constraints/rim_slope_match_out.py
@@ -9,20 +9,6 @@ optional tilt-only projection hook:
 
 where
     phi = (h_out - h_rim) / (r_out - r_rim)
-
-Implementation notes
---------------------
-- Uses a small-slope approximation and only includes shape derivatives along
-  the plane normal (ignores r_hat and radial-distance derivatives), matching
-  the energy counterpart in `modules/energy/rim_slope_match_out.py`.
-- Constraints are applied per-vertex when the disk ring has the same number of
-  vertices as the rim; otherwise the disk tilt uses an arc-length-weighted mean.
-- The constraint gradients are weighted by sqrt(arc-length) to approximate a
-  line-integral weighting.
-- This module provides:
-    * constraint_gradients_array: shape-gradient constraints (positions).
-    * constraint_gradients_tilt_array: tilt-gradient constraints (leaflets).
-    * enforce_tilt_constraint: tilt-only projection to satisfy constraints.
 """
 
 from __future__ import annotations
@@ -32,1576 +18,30 @@ import logging
 import numpy as np
 
 from geometry.entities import Mesh
-from geometry.plane_ops import (
-    fit_plane_normal,
-    order_by_angle,
-    orthonormal_basis_from_normal,
+from modules.constraints.rim_slope_match_diagnostics import (
+    coarse_rim_family_diagnostics,
+    matching_residual_diagnostics,
+    matching_ring_diagnostics,
 )
-from modules.constraints.local_interface_shells import build_local_interface_shell_data
+from modules.constraints.rim_slope_match_gradients import (
+    constraint_gradients_array,
+    constraint_gradients_joint_array,
+    constraint_gradients_rows_array,
+    constraint_gradients_tilt_array,
+    constraint_gradients_tilt_rows_array,
+)
+from modules.constraints.rim_slope_match_params import (
+    _resolve_matching_mode,
+    _use_curved_free_disk_shell2_tilt_continuation,
+    _use_disk_theta_targeting,
+)
+from modules.constraints.rim_slope_match_payload import _build_matching_data
+from modules.constraints.rim_slope_match_utils import (
+    _disk_theta_rows_weights_and_direction,
+    _tilt_target_rows_weights_and_direction,
+)
 
 logger = logging.getLogger("membrane_solver")
-_WARNED_DISK_EQUALS_RIM = False
-
-
-def _resolve_group(global_params, key: str) -> str | None:
-    raw = None if global_params is None else global_params.get(key)
-    if raw is None:
-        return None
-    group = str(raw).strip()
-    return group if group else None
-
-
-def _sanitize_disk_group(
-    *, rim_group: str | None, disk_group: str | None
-) -> str | None:
-    """Disable degenerate disk-group coupling when it equals the rim group."""
-    if rim_group is None or disk_group is None:
-        return disk_group
-    if disk_group != rim_group:
-        return disk_group
-    global _WARNED_DISK_EQUALS_RIM
-    if not _WARNED_DISK_EQUALS_RIM:
-        logger.warning(
-            "rim_slope_match_disk_group matches rim_slope_match_group (%s); "
-            "skipping disk-side coupling to avoid degenerate constraints.",
-            rim_group,
-        )
-        _WARNED_DISK_EQUALS_RIM = True
-    return None
-
-
-def _resolve_center(global_params) -> np.ndarray:
-    center = (
-        None if global_params is None else global_params.get("rim_slope_match_center")
-    )
-    if center is None:
-        center = [0.0, 0.0, 0.0]
-    return np.asarray(center, dtype=float).reshape(3)
-
-
-def _resolve_matching_mode(global_params) -> str:
-    raw = None if global_params is None else global_params.get("rim_slope_match_mode")
-    mode = "pointwise_radial_v1" if raw is None else str(raw).strip().lower()
-    if mode not in {
-        "pointwise_radial_v1",
-        "ring_average_radial_v1",
-        "shared_rim_staggered_v1",
-        "physical_edge_staggered_v1",
-    }:
-        raise ValueError(
-            "rim_slope_match_mode must be 'pointwise_radial_v1' or "
-            "'ring_average_radial_v1' or 'shared_rim_staggered_v1' or "
-            "'physical_edge_staggered_v1'."
-        )
-    return mode
-
-
-def _uses_scaffold_trace_lane(global_params, matching_mode: str) -> bool:
-    """Return whether the current physical-edge lane uses an explicit scaffold trace shell."""
-    if matching_mode != "physical_edge_staggered_v1":
-        return False
-    trace_radius = (
-        None
-        if global_params is None
-        else global_params.get("parity_trace_layer_radius")
-    )
-    outer_shells = (
-        0
-        if global_params is None
-        else int(global_params.get("parity_outer_shells") or 0)
-    )
-    return trace_radius is not None and outer_shells > 0
-
-
-def _uses_outer_shell_tilt_matching(matching_mode: str) -> bool:
-    return matching_mode in {
-        "shared_rim_staggered_v1",
-        "physical_edge_staggered_v1",
-    }
-
-
-def _use_curved_free_disk_shell2_tilt_continuation(global_params) -> bool:
-    """Return whether the shared-rim curved free-disk lane should target shell 2.
-
-    This lane uses the tagged first free shell as a surrogate secant ring for the
-    shape-side rim law. The outer-leaflet tilt continuation should therefore be
-    projected one shell farther, onto the first unconstrained shell matched to
-    that surrogate ring, instead of terminating on the surrogate ring itself.
-    """
-    if global_params is None:
-        return False
-    return (
-        str(global_params.get("rim_slope_match_mode") or "").strip().lower()
-        == "shared_rim_staggered_v1"
-        and str(global_params.get("rim_slope_match_group") or "").strip() == "rim"
-        and str(global_params.get("rim_slope_match_outer_group") or "").strip()
-        == "outer"
-        and str(global_params.get("rim_slope_match_disk_group") or "").strip() == "disk"
-        and str(global_params.get("tilt_thetaB_group_in") or "").strip() == "rim"
-        and bool(global_params.get("tilt_out_exclude_shared_rim_outer_rows"))
-    )
-
-
-def _use_disk_theta_targeting(global_params, matching_mode: str) -> bool:
-    if matching_mode == "physical_edge_staggered_v1":
-        if _uses_scaffold_trace_lane(global_params, matching_mode):
-            return False
-        return True
-    if global_params is None:
-        return False
-    return bool(str(global_params.get("theory_parity_lane") or "").strip())
-
-
-def _tilt_target_rows_weights_and_direction(
-    *,
-    data: dict,
-    positions: np.ndarray,
-    normals: np.ndarray,
-    i: int,
-    matching_mode: str,
-) -> tuple[list[int], list[float], np.ndarray] | None:
-    """Return target rows and tangent radial direction for tilt matching."""
-    r_hat = data["r_hat"]
-    if _uses_outer_shell_tilt_matching(matching_mode):
-        tilt_rows = np.asarray(data.get("tilt_rows", data["outer_rows"]), dtype=int)
-        tilt_idx0 = np.asarray(data.get("tilt_idx0", data["outer_idx0"]), dtype=int)
-        tilt_idx1 = np.asarray(data.get("tilt_idx1", data["outer_idx1"]), dtype=int)
-        tilt_w0 = np.asarray(data.get("tilt_w0", data["outer_w0"]), dtype=float)
-        tilt_w1 = np.asarray(data.get("tilt_w1", data["outer_w1"]), dtype=float)
-        row0 = int(tilt_rows[tilt_idx0[i]])
-        row1 = int(tilt_rows[tilt_idx1[i]])
-        w0 = float(tilt_w0[i])
-        w1 = float(tilt_w1[i])
-        normal = w0 * np.asarray(normals[row0], dtype=float) + w1 * np.asarray(
-            normals[row1], dtype=float
-        )
-        normal_norm = float(np.linalg.norm(normal))
-        if normal_norm < 1.0e-12:
-            return None
-        normal = normal / normal_norm
-        rows = [row0]
-        weights = [w0]
-        if row1 != row0 or abs(w1) > 1.0e-12:
-            rows.append(row1)
-            weights.append(w1)
-    else:
-        row = int(np.asarray(data["rim_rows"], dtype=int)[i])
-        normal = np.asarray(normals[row], dtype=float)
-        rows = [row]
-        weights = [1.0]
-
-    r_dir = np.asarray(r_hat[i], dtype=float) - float(np.dot(r_hat[i], normal)) * normal
-    r_norm = float(np.linalg.norm(r_dir))
-    if r_norm < 1e-12:
-        return None
-    r_dir = r_dir / r_norm
-    return rows, weights, r_dir
-
-
-def _disk_theta_rows_weights_and_direction(
-    *,
-    data: dict,
-    i: int,
-    theta_scalar_active: bool,
-) -> tuple[list[int], list[float], list[np.ndarray]] | None:
-    """Return disk-side rows and directions for the inner scalar-theta law."""
-    disk_rows = data.get("disk_rows")
-    disk_r_hat = data.get("disk_r_hat")
-    if disk_rows is None or disk_r_hat is None:
-        return None
-
-    if theta_scalar_active:
-        if data["local_disk"]:
-            return [int(disk_rows[i])], [1.0], [np.asarray(disk_r_hat[i], dtype=float)]
-        disk_weights = data.get("disk_weights")
-        if disk_weights is None:
-            return None
-        disk_weights = np.asarray(disk_weights, dtype=float)
-        weight_sum = float(np.sum(disk_weights))
-        if weight_sum <= 0.0:
-            return None
-        rows = [int(row_idx) for row_idx in disk_rows]
-        weights = [float(w / weight_sum) for w in disk_weights]
-        dirs = [np.asarray(vec, dtype=float) for vec in disk_r_hat]
-        return rows, weights, dirs
-
-    rows: list[int] = []
-    weights: list[float] = []
-    dirs: list[np.ndarray] = []
-    if data["local_disk"]:
-        rows.append(int(disk_rows[i]))
-        weights.append(1.0)
-        dirs.append(np.asarray(disk_r_hat[i], dtype=float))
-    else:
-        disk_weights = data.get("disk_weights")
-        if disk_weights is None:
-            return None
-        disk_weights = np.asarray(disk_weights, dtype=float)
-        weight_sum = float(np.sum(disk_weights))
-        if weight_sum <= 0.0:
-            return None
-        for row_idx, factor in zip(
-            disk_rows, (disk_weights / weight_sum)[:, None] * disk_r_hat
-        ):
-            rows.append(int(row_idx))
-            weights.append(1.0)
-            dirs.append(np.asarray(factor, dtype=float))
-    return rows, weights, dirs
-
-
-def _resolve_normal(global_params) -> np.ndarray | None:
-    raw = None if global_params is None else global_params.get("rim_slope_match_normal")
-    if raw is None:
-        return None
-    arr = np.asarray(raw, dtype=float).reshape(3)
-    nrm = float(np.linalg.norm(arr))
-    if nrm < 1e-15:
-        return None
-    return arr / nrm
-
-
-def _fit_plane_normal(points: np.ndarray) -> np.ndarray | None:
-    return fit_plane_normal(points)
-
-
-def _orthonormal_basis(normal: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
-    return orthonormal_basis_from_normal(normal)
-
-
-def _collect_group_rows(mesh: Mesh, group: str) -> np.ndarray:
-    cache_key = (mesh._vertex_ids_version, str(group))
-    cache_attr = "_rim_slope_match_group_rows_cache"
-    cached = getattr(mesh, cache_attr, None)
-    if isinstance(cached, dict) and "entries" in cached:
-        if cached.get("key") == cache_key and "rows" in cached:
-            return cached["rows"]
-        entries = cached["entries"]
-        rows = entries.get(cache_key)
-        if rows is not None:
-            return rows
-    else:
-        entries = {}
-        if isinstance(cached, dict) and cached.get("key") == cache_key:
-            return cached["rows"]
-
-    rows: list[int] = []
-    for vid in mesh.vertex_ids:
-        opts = getattr(mesh.vertices[int(vid)], "options", None) or {}
-        if opts.get("rim_slope_match_group") == group:
-            row = mesh.vertex_index_to_row.get(int(vid))
-            if row is not None:
-                rows.append(int(row))
-    out = np.asarray(rows, dtype=int)
-    entries[cache_key] = out
-    if len(entries) > 32:
-        # Keep the cache bounded while preserving recent keys.
-        entries = dict(list(entries.items())[-16:])
-    setattr(mesh, cache_attr, {"entries": entries, "key": cache_key, "rows": out})
-    return out
-
-
-def _order_by_angle(
-    positions: np.ndarray, *, center: np.ndarray, normal: np.ndarray
-) -> np.ndarray:
-    return order_by_angle(positions, center=center, normal=normal)
-
-
-def _arc_length_weights(positions: np.ndarray, order: np.ndarray) -> np.ndarray:
-    n = len(order)
-    if n == 0:
-        return np.zeros(0, dtype=float)
-    pos = positions[order]
-    diffs_next = pos[(np.arange(n) + 1) % n] - pos
-    diffs_prev = pos - pos[(np.arange(n) - 1) % n]
-    l_next = np.linalg.norm(diffs_next, axis=1)
-    l_prev = np.linalg.norm(diffs_prev, axis=1)
-    return 0.5 * (l_next + l_prev)
-
-
-def _arc_length_params(positions: np.ndarray) -> tuple[np.ndarray, float]:
-    n = len(positions)
-    if n == 0:
-        return np.zeros(0, dtype=float), 0.0
-    diffs = positions[(np.arange(n) + 1) % n] - positions
-    seg = np.linalg.norm(diffs, axis=1)
-    total = float(np.sum(seg))
-    if total <= 0.0:
-        return np.zeros(n, dtype=float), 0.0
-    s = np.concatenate(([0.0], np.cumsum(seg[:-1], dtype=float)))
-    s /= total
-    return s, total
-
-
-def _interp_ring_positions(
-    positions: np.ndarray, s_targets: np.ndarray
-) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray] | None:
-    n = len(positions)
-    if n == 0 or s_targets.size == 0:
-        return None
-    s_out, total = _arc_length_params(positions)
-    if total <= 0.0 or s_out.size < 2:
-        return None
-    idx1 = np.searchsorted(s_out, s_targets, side="right") % n
-    idx0 = (idx1 - 1) % n
-    s0 = s_out[idx0]
-    s1 = s_out[idx1]
-    s1_adj = s1.copy()
-    s_targets_adj = s_targets.copy()
-    wrap = s1_adj <= s0
-    s1_adj[wrap] += 1.0
-    s_targets_adj = np.where(s_targets_adj < s0, s_targets_adj + 1.0, s_targets_adj)
-    denom = s1_adj - s0
-    t = np.zeros_like(s_targets_adj)
-    mask = denom > 1e-12
-    t[mask] = (s_targets_adj[mask] - s0[mask]) / denom[mask]
-    w1 = t
-    w0 = 1.0 - t
-    interp = positions[idx0] * w0[:, None] + positions[idx1] * w1[:, None]
-    return interp, idx0, idx1, w0, w1
-
-
-def _build_matching_data(mesh: Mesh, global_params, positions: np.ndarray):
-    matching_mode = _resolve_matching_mode(global_params)
-    group = _resolve_group(global_params, "rim_slope_match_group")
-    outer_group = _resolve_group(global_params, "rim_slope_match_outer_group")
-    disk_group = _resolve_group(global_params, "rim_slope_match_disk_group")
-    if matching_mode != "physical_edge_staggered_v1":
-        disk_group = _sanitize_disk_group(rim_group=group, disk_group=disk_group)
-    theta_param = None
-    if global_params is not None:
-        theta_param = global_params.get("rim_slope_match_thetaB_param")
-    if matching_mode == "physical_edge_staggered_v1":
-        interface_group = disk_group or group
-        if interface_group is None:
-            return None
-    elif group is None or outer_group is None:
-        return None
-    center = _resolve_center(global_params)
-    raw_normal = _resolve_normal(global_params)
-    normal_token = (
-        None
-        if raw_normal is None
-        else tuple(np.asarray(raw_normal, dtype=float).reshape(3).tolist())
-    )
-    theta_token = None
-    if (
-        theta_param is not None
-        and global_params is not None
-        and not _uses_scaffold_trace_lane(global_params, matching_mode)
-    ):
-        theta_token = float(global_params.get(str(theta_param)) or 0.0)
-    trace_layer_radius = (
-        None
-        if global_params is None
-        else global_params.get("parity_trace_layer_radius")
-    )
-    cache_key = (
-        int(mesh._version),
-        int(mesh._vertex_ids_version),
-        str(matching_mode),
-        str(group),
-        str(outer_group),
-        None if disk_group is None else str(disk_group),
-        None if theta_param is None else str(theta_param),
-        theta_token,
-        None if trace_layer_radius is None else float(trace_layer_radius),
-        tuple(center.tolist()),
-        normal_token,
-        id(positions),
-    )
-    cache_attr = "_rim_slope_match_data_cache"
-    if mesh._geometry_cache_active(positions):
-        cached = getattr(mesh, cache_attr, None)
-        if isinstance(cached, dict) and "entries" in cached:
-            value = cached["entries"].get(cache_key)
-            if value is not None:
-                return value
-        elif cached is not None and cached.get("key") == cache_key:
-            return cached.get("value")
-
-    normal = raw_normal
-    if matching_mode == "physical_edge_staggered_v1":
-        try:
-            local_shells = build_local_interface_shell_data(
-                mesh, positions=positions, group=str(interface_group)
-            )
-        except AssertionError:
-            return None
-        scaffold_outer_shells = (
-            0
-            if global_params is None
-            else int(global_params.get("parity_outer_shells") or 0)
-        )
-        scaffold_trace_tilt = (
-            trace_layer_radius is not None and scaffold_outer_shells > 0
-        )
-        if trace_layer_radius is None:
-            rim_rows = np.asarray(local_shells.disk_rows, dtype=int)
-            outer_rows = np.asarray(local_shells.rim_rows_for_disk, dtype=int)
-            tilt_rows = outer_rows.copy()
-        else:
-            rim_rows = np.asarray(local_shells.disk_rows, dtype=int)
-            if scaffold_trace_tilt:
-                outer_rows = np.asarray(local_shells.rim_rows_for_disk, dtype=int)
-                tilt_rows = outer_rows.copy()
-            else:
-                outer_rows = np.asarray(local_shells.outer_rows_for_disk, dtype=int)
-                tilt_rows = outer_rows.copy()
-        if rim_rows.size == 0 or outer_rows.size == 0:
-            return None
-        if normal is None:
-            normal = _fit_plane_normal(positions[rim_rows])
-        if normal is None:
-            normal = np.array([0.0, 0.0, 1.0], dtype=float)
-        rim_pos = positions[rim_rows]
-        outer_pos = positions[outer_rows]
-        tilt_pos = positions[tilt_rows]
-    else:
-        rim_rows = _collect_group_rows(mesh, group)
-        outer_rows = _collect_group_rows(mesh, outer_group)
-        if rim_rows.size == 0 or outer_rows.size == 0:
-            return None
-        if normal is None:
-            normal = _fit_plane_normal(positions[rim_rows])
-        if normal is None:
-            normal = np.array([0.0, 0.0, 1.0], dtype=float)
-
-        rim_pos = positions[rim_rows]
-        outer_pos = positions[outer_rows]
-
-        rim_order = _order_by_angle(rim_pos, center=center, normal=normal)
-        outer_order = _order_by_angle(outer_pos, center=center, normal=normal)
-
-        rim_rows = rim_rows[rim_order]
-        outer_rows = outer_rows[outer_order]
-        rim_pos = positions[rim_rows]
-        outer_pos = positions[outer_rows]
-        tilt_rows = outer_rows.copy()
-        tilt_pos = outer_pos
-        if (
-            matching_mode == "shared_rim_staggered_v1"
-            and _use_curved_free_disk_shell2_tilt_continuation(global_params)
-        ):
-            try:
-                local_shells = build_local_interface_shell_data(
-                    mesh, positions=positions, group=str(group)
-                )
-            except AssertionError:
-                local_shells = None
-            if local_shells is not None:
-                local_rim_rows = np.asarray(local_shells.rim_rows, dtype=int)
-                local_outer_for_rim = np.asarray(
-                    local_shells.outer_rows_for_rim, dtype=int
-                )
-                row_to_target = {
-                    int(row): int(target)
-                    for row, target in zip(local_rim_rows, local_outer_for_rim)
-                }
-                matched_outer = np.asarray(
-                    [row_to_target.get(int(row), -1) for row in outer_rows],
-                    dtype=int,
-                )
-                if matched_outer.size == outer_rows.size and np.all(matched_outer >= 0):
-                    tilt_rows = matched_outer
-                    tilt_pos = positions[tilt_rows]
-    outer_idx0 = np.arange(len(rim_rows), dtype=int)
-    outer_idx1 = np.arange(len(rim_rows), dtype=int)
-    outer_w0 = np.ones(len(rim_rows), dtype=float)
-    outer_w1 = np.zeros(len(rim_rows), dtype=float)
-    if rim_rows.size != outer_rows.size:
-        s_rim, total_rim = _arc_length_params(rim_pos)
-        if total_rim <= 0.0:
-            return None
-        interp = _interp_ring_positions(outer_pos, s_rim)
-        if interp is None:
-            return None
-        outer_pos, outer_idx0, outer_idx1, outer_w0, outer_w1 = interp
-    tilt_idx0 = np.arange(len(rim_rows), dtype=int)
-    tilt_idx1 = np.arange(len(rim_rows), dtype=int)
-    tilt_w0 = np.ones(len(rim_rows), dtype=float)
-    tilt_w1 = np.zeros(len(rim_rows), dtype=float)
-    if rim_rows.size != tilt_rows.size:
-        s_rim, total_rim = _arc_length_params(rim_pos)
-        if total_rim <= 0.0:
-            return None
-        interp_tilt = _interp_ring_positions(tilt_pos, s_rim)
-        if interp_tilt is None:
-            return None
-        _tilt_pos_interp, tilt_idx0, tilt_idx1, tilt_w0, tilt_w1 = interp_tilt
-
-    r_vec = rim_pos - center[None, :]
-    r_vec = r_vec - np.einsum("ij,j->i", r_vec, normal)[:, None] * normal[None, :]
-    r_len = np.linalg.norm(r_vec, axis=1)
-    good = r_len > 1e-12
-    if not np.any(good):
-        return None
-
-    r_hat = np.zeros_like(r_vec)
-    r_hat[good] = r_vec[good] / r_len[good][:, None]
-
-    h_rim = np.einsum("ij,j->i", rim_pos - center[None, :], normal)
-    h_out = np.einsum("ij,j->i", outer_pos - center[None, :], normal)
-    r_out_vec = outer_pos - center[None, :]
-    r_out_vec = (
-        r_out_vec - np.einsum("ij,j->i", r_out_vec, normal)[:, None] * normal[None, :]
-    )
-    r_out = np.linalg.norm(r_out_vec, axis=1)
-    dr = r_out - r_len
-    inv_dr = np.zeros_like(dr)
-    valid = good & (np.abs(dr) > 1e-8)
-    inv_dr[valid] = 1.0 / dr[valid]
-
-    phi = np.zeros_like(dr)
-    phi[valid] = (h_out[valid] - h_rim[valid]) * inv_dr[valid]
-
-    weights = _arc_length_weights(rim_pos, np.arange(len(rim_rows)))
-    weights = np.where(valid, weights, 0.0)
-    weight_sqrt = np.sqrt(weights)
-
-    disk_rows = None
-    disk_r_hat = None
-    disk_weights = None
-    local_disk = False
-    theta_scalar = theta_token
-    if matching_mode == "physical_edge_staggered_v1":
-        disk_rows = rim_rows.copy()
-        disk_r_hat = r_hat.copy()
-        disk_weights = None
-        local_disk = True
-    elif disk_group is not None:
-        disk_rows = _collect_group_rows(mesh, disk_group)
-        if disk_rows.size:
-            disk_pos = positions[disk_rows]
-            disk_order = _order_by_angle(disk_pos, center=center, normal=normal)
-            disk_rows = disk_rows[disk_order]
-            disk_pos = positions[disk_rows]
-            r_vec_disk = disk_pos - center[None, :]
-            r_vec_disk = (
-                r_vec_disk
-                - np.einsum("ij,j->i", r_vec_disk, normal)[:, None] * normal[None, :]
-            )
-            r_len_disk = np.linalg.norm(r_vec_disk, axis=1)
-            disk_r_hat = np.zeros_like(r_vec_disk)
-            good_disk = r_len_disk > 1e-12
-            disk_r_hat[good_disk] = (
-                r_vec_disk[good_disk] / r_len_disk[good_disk][:, None]
-            )
-            if disk_rows.size == rim_rows.size:
-                local_disk = True
-            else:
-                disk_weights = _arc_length_weights(disk_pos, np.arange(len(disk_rows)))
-                disk_weights = np.where(good_disk, disk_weights, 0.0)
-
-    result = {
-        "rim_rows": rim_rows,
-        "outer_rows": outer_rows,
-        "outer_idx0": outer_idx0,
-        "outer_idx1": outer_idx1,
-        "outer_w0": outer_w0,
-        "outer_w1": outer_w1,
-        "tilt_rows": tilt_rows,
-        "tilt_idx0": tilt_idx0,
-        "tilt_idx1": tilt_idx1,
-        "tilt_w0": tilt_w0,
-        "tilt_w1": tilt_w1,
-        "disk_rows": disk_rows,
-        "r_hat": r_hat,
-        "disk_r_hat": disk_r_hat,
-        "weight_sqrt": weight_sqrt,
-        "inv_dr": inv_dr,
-        "phi": phi,
-        "valid": valid,
-        "local_disk": local_disk,
-        "disk_weights": disk_weights,
-        "normal": normal,
-        "theta_scalar": theta_scalar,
-        "matching_mode": matching_mode,
-        "construction_mode": (
-            "physical_edge_local_shell"
-            if matching_mode == "physical_edge_staggered_v1"
-            else "legacy_tagged_rim_shell"
-        ),
-    }
-    if mesh._geometry_cache_active(positions):
-        cached = getattr(mesh, cache_attr, None)
-        entries = cached.get("entries", {}) if isinstance(cached, dict) else {}
-        entries[cache_key] = result
-        if len(entries) > 16:
-            entries = dict(list(entries.items())[-8:])
-        setattr(mesh, cache_attr, {"entries": entries})
-    return result
-
-
-def matching_ring_diagnostics(mesh: Mesh, global_params, positions: np.ndarray) -> dict:
-    """Report tagged ring candidates seen by the rim-matching constraint path."""
-    matching_mode = _resolve_matching_mode(global_params)
-    group = _resolve_group(global_params, "rim_slope_match_group")
-    outer_group = _resolve_group(global_params, "rim_slope_match_outer_group")
-    disk_group = _resolve_group(global_params, "rim_slope_match_disk_group")
-    if matching_mode != "physical_edge_staggered_v1":
-        disk_group = _sanitize_disk_group(rim_group=group, disk_group=disk_group)
-    center = _resolve_center(global_params)
-    normal = _resolve_normal(global_params)
-
-    out: dict[str, object] = {
-        "available": False,
-        "reason": "missing_groups",
-        "rim_group": group,
-        "outer_group": outer_group,
-        "disk_group": disk_group,
-        "outer_source": "tagged_group",
-        "matching_mode": matching_mode,
-    }
-    if matching_mode == "physical_edge_staggered_v1":
-        interface_group = disk_group or group
-        out["reason"] = "missing_interface_group"
-        out["outer_source"] = "local_interface_shell"
-        if interface_group is None:
-            return out
-        try:
-            local_shells = build_local_interface_shell_data(
-                mesh, positions=positions, group=str(interface_group)
-            )
-        except AssertionError as exc:
-            out["reason"] = str(exc)
-            return out
-        out.update(
-            {
-                "available": True,
-                "reason": "ok",
-                "construction_mode": "physical_edge_local_shell",
-                "rim_count": int(local_shells.disk_rows.size),
-                "outer_count": int(local_shells.rim_rows.size),
-                "disk_count": int(local_shells.disk_rows.size),
-                "rim_radius": float(local_shells.disk_radius),
-                "outer_radius": float(local_shells.rim_radius),
-                "disk_radius": float(local_shells.disk_radius),
-            }
-        )
-        return out
-    if group is None or outer_group is None:
-        return out
-
-    rim_rows = _collect_group_rows(mesh, group)
-    outer_rows = _collect_group_rows(mesh, outer_group)
-    disk_rows = (
-        np.zeros(0, dtype=int)
-        if disk_group is None
-        else _collect_group_rows(mesh, disk_group)
-    )
-    if rim_rows.size == 0:
-        out["reason"] = "missing_rim_group_rows"
-    elif outer_rows.size == 0:
-        out["reason"] = "missing_outer_group_rows"
-    else:
-        out["available"] = True
-        out["reason"] = "ok"
-
-    if normal is None and rim_rows.size > 0:
-        normal = _fit_plane_normal(positions[rim_rows])
-    if normal is None:
-        normal = np.array([0.0, 0.0, 1.0], dtype=float)
-
-    rel_all = positions - center[None, :]
-    rel_all = rel_all - np.einsum("ij,j->i", rel_all, normal)[:, None] * normal[None, :]
-    r_all = np.linalg.norm(rel_all, axis=1)
-
-    def _median_radius(rows: np.ndarray) -> float:
-        if rows.size == 0:
-            return float("nan")
-        return float(np.median(r_all[rows]))
-
-    out["rim_count"] = int(rim_rows.size)
-    out["outer_count"] = int(outer_rows.size)
-    out["disk_count"] = int(disk_rows.size)
-    out["rim_radius"] = _median_radius(rim_rows)
-    out["outer_radius"] = _median_radius(outer_rows)
-    out["disk_radius"] = _median_radius(disk_rows)
-    out["construction_mode"] = "legacy_tagged_rim_shell"
-    return out
-
-
-def coarse_rim_family_diagnostics(
-    mesh: Mesh, global_params, positions: np.ndarray
-) -> dict[str, object]:
-    """Describe the coarse tagged rim family and its immediate radial neighbors."""
-    group = _resolve_group(global_params, "rim_slope_match_group")
-    center = _resolve_center(global_params)
-    normal = _resolve_normal(global_params)
-    out: dict[str, object] = {
-        "available": False,
-        "reason": "missing_rim_group",
-        "rim_group": group,
-    }
-    if group is None:
-        return out
-
-    rim_rows = _collect_group_rows(mesh, group)
-    if rim_rows.size == 0:
-        out["reason"] = "missing_rim_group_rows"
-        return out
-
-    if normal is None:
-        normal = _fit_plane_normal(positions[rim_rows])
-    if normal is None:
-        normal = np.array([0.0, 0.0, 1.0], dtype=float)
-
-    rel_all = positions - center[None, :]
-    rel_all = rel_all - np.einsum("ij,j->i", rel_all, normal)[:, None] * normal[None, :]
-    r_all = np.linalg.norm(rel_all, axis=1)
-    rim_radii = r_all[rim_rows]
-    rim_radius_unique, rim_radius_counts = np.unique(
-        np.round(rim_radii, 12), return_counts=True
-    )
-
-    rim_vids = {int(mesh.vertex_ids[row]) for row in rim_rows.tolist()}
-    neighbor_rows: set[int] = set()
-    for edge in mesh.edges.values():
-        tail = int(edge.tail_index)
-        head = int(edge.head_index)
-        if tail in rim_vids and head not in rim_vids:
-            row = mesh.vertex_index_to_row.get(head)
-            if row is not None:
-                neighbor_rows.add(int(row))
-        if head in rim_vids and tail not in rim_vids:
-            row = mesh.vertex_index_to_row.get(tail)
-            if row is not None:
-                neighbor_rows.add(int(row))
-
-    neighbor_rows_arr = np.asarray(sorted(neighbor_rows), dtype=int)
-    inner_rows = np.zeros(0, dtype=int)
-    outer_rows = np.zeros(0, dtype=int)
-    if neighbor_rows_arr.size:
-        tol = max(1.0e-9, 1.0e-5 * max(1.0, float(np.max(rim_radii))))
-        rim_r_min = float(np.min(rim_radii))
-        rim_r_max = float(np.max(rim_radii))
-        neighbor_r = r_all[neighbor_rows_arr]
-        inner_rows = neighbor_rows_arr[neighbor_r < (rim_r_min - tol)]
-        outer_rows = neighbor_rows_arr[neighbor_r > (rim_r_max + tol)]
-
-    def _radius_block(rows: np.ndarray) -> dict[str, object]:
-        if rows.size == 0:
-            return {
-                "count": 0,
-                "radius_min": float("nan"),
-                "radius_max": float("nan"),
-                "radius_unique": [],
-            }
-        vals = r_all[rows]
-        unique_vals = np.unique(np.round(vals, 12))
-        return {
-            "count": int(rows.size),
-            "radius_min": float(np.min(vals)),
-            "radius_max": float(np.max(vals)),
-            "radius_unique": [float(v) for v in unique_vals.tolist()],
-        }
-
-    out.update(
-        {
-            "available": True,
-            "reason": "ok",
-            "rim_count": int(rim_rows.size),
-            "rim_radius_min": float(np.min(rim_radii)),
-            "rim_radius_max": float(np.max(rim_radii)),
-            "rim_radius_unique": [float(v) for v in rim_radius_unique.tolist()],
-            "rim_radius_unique_counts": [int(v) for v in rim_radius_counts.tolist()],
-            "inner_neighbor": _radius_block(inner_rows),
-            "outer_neighbor": _radius_block(outer_rows),
-        }
-    )
-    return out
-
-
-def _residual_summary(values: np.ndarray, angles: np.ndarray) -> dict[str, float]:
-    """Summarize residual structure with bulk stats and first azimuthal mode."""
-    if values.size == 0:
-        return {
-            "mean": float("nan"),
-            "median": float("nan"),
-            "std": float("nan"),
-            "max_abs": float("nan"),
-            "first_mode_cos": float("nan"),
-            "first_mode_sin": float("nan"),
-            "first_mode_amplitude": float("nan"),
-        }
-    mean = float(np.mean(values))
-    median = float(np.median(values))
-    std = float(np.std(values))
-    max_abs = float(np.max(np.abs(values)))
-    cos1 = float(np.mean(values * np.cos(angles)))
-    sin1 = float(np.mean(values * np.sin(angles)))
-    amp1 = float(np.hypot(cos1, sin1))
-    return {
-        "mean": mean,
-        "median": median,
-        "std": std,
-        "max_abs": max_abs,
-        "first_mode_cos": cos1,
-        "first_mode_sin": sin1,
-        "first_mode_amplitude": amp1,
-    }
-
-
-def matching_residual_diagnostics(
-    mesh: Mesh, global_params, positions: np.ndarray
-) -> dict[str, object]:
-    """Report residuals for the active coarse rim-matching law without enforcing it."""
-    ring_diag = matching_ring_diagnostics(mesh, global_params, positions)
-    data = _build_matching_data(mesh, global_params, positions)
-    out: dict[str, object] = {
-        "available": False,
-        "reason": str(ring_diag.get("reason", "missing_matching_data")),
-        "phi_estimator": "tagged_group_secant_dr",
-        "matching_mode": _resolve_matching_mode(global_params),
-        "disk_theta_source": "unavailable",
-    }
-    if data is None:
-        return out
-
-    rim_rows = data["rim_rows"]
-    disk_rows = data["disk_rows"]
-    outer_rows = data["outer_rows"]
-    r_hat = data["r_hat"]
-    disk_r_hat = data["disk_r_hat"]
-    phi = data["phi"]
-    valid = np.asarray(data["valid"], dtype=bool)
-    local_disk = bool(data["local_disk"])
-    disk_weights = data["disk_weights"]
-    normal = np.asarray(data["normal"], dtype=float)
-    theta_scalar = data["theta_scalar"]
-    center = _resolve_center(global_params)
-
-    rim_pos = positions[rim_rows]
-    u, v = _orthonormal_basis(normal)
-    rel = rim_pos - center[None, :]
-    rel_plane = rel - np.einsum("ij,j->i", rel, normal)[:, None] * normal[None, :]
-    rim_angles = np.arctan2(rel_plane @ v, rel_plane @ u)
-
-    normals = mesh.vertex_normals(positions=positions)
-    matching_mode = str(out["matching_mode"])
-    r_dir = np.zeros_like(r_hat)
-    good_dir = np.zeros(r_hat.shape[0], dtype=bool)
-    tilt_rows0 = np.asarray(data["rim_rows"], dtype=int)
-    tilt_rows1 = np.asarray(data["rim_rows"], dtype=int)
-    tilt_w0 = np.ones(r_hat.shape[0], dtype=float)
-    tilt_w1 = np.zeros(r_hat.shape[0], dtype=float)
-    if matching_mode == "shared_rim_staggered_v1":
-        outer_rows = np.asarray(data["outer_rows"], dtype=int)
-        tilt_rows0 = outer_rows[np.asarray(data["outer_idx0"], dtype=int)]
-        tilt_rows1 = outer_rows[np.asarray(data["outer_idx1"], dtype=int)]
-        tilt_w0 = np.asarray(data["outer_w0"], dtype=float)
-        tilt_w1 = np.asarray(data["outer_w1"], dtype=float)
-    elif matching_mode == "physical_edge_staggered_v1":
-        tilt_rows = np.asarray(data.get("tilt_rows", data["outer_rows"]), dtype=int)
-        tilt_rows0 = tilt_rows[
-            np.asarray(data.get("tilt_idx0", data["outer_idx0"]), dtype=int)
-        ]
-        tilt_rows1 = tilt_rows[
-            np.asarray(data.get("tilt_idx1", data["outer_idx1"]), dtype=int)
-        ]
-        tilt_w0 = np.asarray(data.get("tilt_w0", data["outer_w0"]), dtype=float)
-        tilt_w1 = np.asarray(data.get("tilt_w1", data["outer_w1"]), dtype=float)
-    for i in range(r_hat.shape[0]):
-        n = tilt_w0[i] * normals[tilt_rows0[i]] + tilt_w1[i] * normals[tilt_rows1[i]]
-        n_norm = float(np.linalg.norm(n))
-        if n_norm < 1e-12:
-            continue
-        n = n / n_norm
-        tangent_dir = r_hat[i] - float(np.dot(r_hat[i], n)) * n
-        tangent_norm = float(np.linalg.norm(tangent_dir))
-        if tangent_norm < 1e-12:
-            continue
-        r_dir[i] = tangent_dir / tangent_norm
-        good_dir[i] = True
-
-    valid = valid & good_dir
-    out.update(
-        {
-            "available": bool(np.any(valid)),
-            "reason": "ok" if bool(np.any(valid)) else "no_valid_rows",
-            "sample_count": int(rim_rows.size),
-            "valid_count": int(np.count_nonzero(valid)),
-            "rim_count": int(rim_rows.size),
-            "disk_count": 0 if disk_rows is None else int(disk_rows.size),
-            "local_disk": bool(local_disk),
-            "theta_param_used": bool(theta_scalar is not None),
-        }
-    )
-    if not np.any(valid):
-        return out
-
-    phi_valid = phi[valid]
-    angles_valid = rim_angles[valid]
-    tilt_out_rad = (
-        tilt_w0 * np.einsum("ij,ij->i", mesh.tilts_out_view()[tilt_rows0], r_dir)
-        + tilt_w1 * np.einsum("ij,ij->i", mesh.tilts_out_view()[tilt_rows1], r_dir)
-    )[valid]
-    outer_residual = tilt_out_rad - phi_valid
-
-    out["phi_mean"] = float(np.mean(phi_valid))
-    out["phi_median"] = float(np.median(phi_valid))
-    out["phi_std"] = float(np.std(phi_valid))
-    out["phi_max_abs"] = float(np.max(np.abs(phi_valid)))
-    out["outer_residual"] = _residual_summary(outer_residual, angles_valid)
-
-    if disk_rows is None or disk_r_hat is None:
-        out["disk_theta_source"] = "unavailable"
-        out["inner_residual_available"] = False
-        return out
-
-    if theta_scalar is not None:
-        theta_disk = np.full(phi.shape, float(theta_scalar), dtype=float)
-        out["disk_theta_source"] = "global_param"
-    elif local_disk:
-        theta_disk = np.einsum("ij,ij->i", mesh.tilts_in_view()[disk_rows], disk_r_hat)
-        out["disk_theta_source"] = "local_disk_rows"
-    else:
-        if disk_weights is None:
-            out["disk_theta_source"] = "missing_disk_weights"
-            out["inner_residual_available"] = False
-            return out
-        weight_sum = float(np.sum(disk_weights))
-        if weight_sum <= 0.0:
-            out["disk_theta_source"] = "degenerate_disk_weights"
-            out["inner_residual_available"] = False
-            return out
-        theta_disk_scalar = float(
-            np.sum(
-                np.asarray(disk_weights, dtype=float)
-                * np.einsum("ij,ij->i", mesh.tilts_in_view()[disk_rows], disk_r_hat)
-            )
-            / weight_sum
-        )
-        theta_disk = np.full(phi.shape, theta_disk_scalar, dtype=float)
-        out["disk_theta_source"] = "weighted_disk_mean"
-
-    if theta_scalar is not None and _use_disk_theta_targeting(
-        global_params, matching_mode
-    ):
-        if local_disk:
-            tilt_in_rad = np.einsum(
-                "ij,ij->i", mesh.tilts_in_view()[disk_rows], disk_r_hat
-            )[valid]
-        else:
-            weight_sum = (
-                float(np.sum(disk_weights)) if disk_weights is not None else 0.0
-            )
-            if weight_sum <= 0.0:
-                out["disk_theta_source"] = "degenerate_disk_weights"
-                out["inner_residual_available"] = False
-                return out
-            tilt_in_scalar = float(
-                np.sum(
-                    np.asarray(disk_weights, dtype=float)
-                    * np.einsum("ij,ij->i", mesh.tilts_in_view()[disk_rows], disk_r_hat)
-                )
-                / weight_sum
-            )
-            tilt_in_rad = np.full(np.count_nonzero(valid), tilt_in_scalar, dtype=float)
-    else:
-        tilt_in_rad = (
-            tilt_w0 * np.einsum("ij,ij->i", mesh.tilts_in_view()[tilt_rows0], r_dir)
-            + tilt_w1 * np.einsum("ij,ij->i", mesh.tilts_in_view()[tilt_rows1], r_dir)
-        )[valid]
-    inner_target = theta_disk[valid] - phi_valid
-    inner_residual = tilt_in_rad - inner_target
-    out["inner_residual_available"] = True
-    out["theta_disk_mean"] = float(np.mean(theta_disk[valid]))
-    out["theta_disk_median"] = float(np.median(theta_disk[valid]))
-    out["inner_residual"] = _residual_summary(inner_residual, angles_valid)
-    return out
-
-
-def constraint_gradients_joint_array(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: dict[int, int],
-    tilts_in: np.ndarray | None = None,
-    tilts_out: np.ndarray | None = None,
-) -> (
-    list[
-        tuple[
-            tuple[np.ndarray, np.ndarray] | None,  # shape
-            tuple[np.ndarray, np.ndarray] | None,  # tilt_in
-            tuple[np.ndarray, np.ndarray] | None,  # tilt_out
-        ]
-    ]
-    | None
-):
-    """Return paired shape/tilt gradients for joint KKT projection."""
-    _ = tilts_in, tilts_out
-    data = _build_matching_data(mesh, global_params, positions)
-    if data is None:
-        return None
-
-    rim_rows = data["rim_rows"]
-    outer_rows = data["outer_rows"]
-    outer_idx0 = data["outer_idx0"]
-    outer_idx1 = data["outer_idx1"]
-    outer_w0 = data["outer_w0"]
-    outer_w1 = data["outer_w1"]
-    disk_rows = data["disk_rows"]
-    disk_r_hat = data["disk_r_hat"]
-    weight_sqrt = data["weight_sqrt"]
-    inv_dr = data["inv_dr"]
-    valid = data["valid"]
-    normal = data["normal"]
-    matching_mode = _resolve_matching_mode(global_params)
-    theta_scalar = data["theta_scalar"]
-
-    normals = mesh.vertex_normals(positions=positions)
-
-    if matching_mode != "physical_edge_staggered_v1" and not _use_disk_theta_targeting(
-        global_params, matching_mode
-    ):
-        constraints: list[
-            tuple[
-                tuple[np.ndarray, np.ndarray] | None,
-                tuple[np.ndarray, np.ndarray] | None,
-                tuple[np.ndarray, np.ndarray] | None,
-            ]
-        ] = []
-        agg_shape_outer_rows: list[int] = []
-        agg_shape_outer_vecs: list[np.ndarray] = []
-        agg_out_rows: list[int] = []
-        agg_out_vecs: list[np.ndarray] = []
-        agg_shape_inner_rows: list[int] = []
-        agg_shape_inner_vecs: list[np.ndarray] = []
-        agg_in_rows: list[int] = []
-        agg_in_vecs: list[np.ndarray] = []
-
-        for i, ok in enumerate(valid):
-            if not ok or weight_sqrt[i] == 0.0:
-                continue
-            coeff_shape = weight_sqrt[i] * inv_dr[i]
-            coeff_tilt = weight_sqrt[i]
-
-            target = _tilt_target_rows_weights_and_direction(
-                data=data,
-                positions=positions,
-                normals=normals,
-                i=i,
-                matching_mode=matching_mode,
-            )
-            if target is None:
-                continue
-            target_rows, target_weights, r_dir = target
-
-            out0 = outer_rows[outer_idx0[i]]
-            out1 = outer_rows[outer_idx1[i]]
-
-            rows_shape = [int(rim_rows[i]), int(out0)]
-            vecs_shape_outer = [
-                coeff_shape * normal,
-                -coeff_shape * outer_w0[i] * normal,
-            ]
-            vecs_shape_inner = [
-                -coeff_shape * normal,
-                coeff_shape * outer_w0[i] * normal,
-            ]
-            if out1 != out0 or outer_w1[i] != 0.0:
-                rows_shape.append(int(out1))
-                vecs_shape_outer.append(-coeff_shape * outer_w1[i] * normal)
-                vecs_shape_inner.append(coeff_shape * outer_w1[i] * normal)
-
-            rows_tilt = [int(row) for row in target_rows]
-            vecs_tilt = [coeff_tilt * float(w) * r_dir for w in target_weights]
-            out_part = (
-                np.asarray(rows_tilt, dtype=int),
-                np.asarray(vecs_tilt, dtype=float),
-            )
-            shape_part_outer = (
-                np.asarray(rows_shape, dtype=int),
-                np.asarray(vecs_shape_outer, dtype=float),
-            )
-            if matching_mode == "ring_average_radial_v1":
-                agg_shape_outer_rows.extend(shape_part_outer[0].tolist())
-                agg_shape_outer_vecs.extend(shape_part_outer[1])
-                agg_out_rows.extend(out_part[0].tolist())
-                agg_out_vecs.extend(out_part[1])
-            else:
-                constraints.append((shape_part_outer, None, out_part))
-
-            if disk_rows is None or disk_r_hat is None:
-                continue
-
-            rows_in = [int(row) for row in target_rows]
-            vecs_in = [coeff_tilt * float(w) * r_dir for w in target_weights]
-            if theta_scalar is None:
-                if data["local_disk"]:
-                    rows_in.append(int(disk_rows[i]))
-                    vecs_in.append(-coeff_tilt * disk_r_hat[i])
-                elif data["disk_weights"] is not None:
-                    disk_weights = np.asarray(data["disk_weights"], dtype=float)
-                    weight_sum = float(np.sum(disk_weights))
-                    if weight_sum > 0.0:
-                        factors = (disk_weights / weight_sum)[:, None] * disk_r_hat
-                        for row_idx, factor in zip(disk_rows, factors):
-                            rows_in.append(int(row_idx))
-                            vecs_in.append(-coeff_tilt * factor)
-
-            shape_part_inner = (
-                np.asarray(rows_shape, dtype=int),
-                np.asarray(vecs_shape_inner, dtype=float),
-            )
-            in_part = (
-                np.asarray(rows_in, dtype=int),
-                np.asarray(vecs_in, dtype=float),
-            )
-            if matching_mode == "ring_average_radial_v1":
-                agg_shape_inner_rows.extend(shape_part_inner[0].tolist())
-                agg_shape_inner_vecs.extend(shape_part_inner[1])
-                agg_in_rows.extend(in_part[0].tolist())
-                agg_in_vecs.extend(in_part[1])
-            else:
-                constraints.append((shape_part_inner, in_part, None))
-
-        if matching_mode == "ring_average_radial_v1":
-            if agg_shape_outer_rows or agg_out_rows:
-                constraints.append(
-                    (
-                        (
-                            np.asarray(agg_shape_outer_rows, dtype=int),
-                            np.asarray(agg_shape_outer_vecs, dtype=float),
-                        ),
-                        None,
-                        (
-                            np.asarray(agg_out_rows, dtype=int),
-                            np.asarray(agg_out_vecs, dtype=float),
-                        ),
-                    )
-                )
-            if agg_shape_inner_rows or agg_in_rows:
-                constraints.append(
-                    (
-                        (
-                            np.asarray(agg_shape_inner_rows, dtype=int),
-                            np.asarray(agg_shape_inner_vecs, dtype=float),
-                        ),
-                        (
-                            np.asarray(agg_in_rows, dtype=int),
-                            np.asarray(agg_in_vecs, dtype=float),
-                        ),
-                        None,
-                    )
-                )
-
-        return constraints or None
-
-    constraints: list[
-        tuple[
-            tuple[np.ndarray, np.ndarray] | None,
-            tuple[np.ndarray, np.ndarray] | None,
-            tuple[np.ndarray, np.ndarray] | None,
-        ]
-    ] = []
-
-    for i, ok in enumerate(valid):
-        if not ok or weight_sqrt[i] == 0.0:
-            continue
-        coeff_s = weight_sqrt[i] * inv_dr[i]
-        coeff_t = weight_sqrt[i]
-
-        target = _tilt_target_rows_weights_and_direction(
-            data=data,
-            positions=positions,
-            normals=normals,
-            i=i,
-            matching_mode=matching_mode,
-        )
-        if target is None:
-            continue
-        target_rows, target_weights, r_dir = target
-
-        # 1. Outer constraint: t_out . r_dir - phi = 0
-        # Partials: dC/dt_out = r_dir, dC/dx = -d(phi)/dx
-
-        # d(phi)/dx component (shape)
-        # phi = (h_out - h_rim) / dr
-        # d(phi)/dx_rim = -normal / dr
-        # d(phi)/dx_out = +normal / dr
-        out0 = outer_rows[outer_idx0[i]]
-        out1 = outer_rows[outer_idx1[i]]
-
-        rows_s = [int(rim_rows[i]), int(out0)]
-        # dC/dx = -d(phi)/dx => rim component is +normal/dr, out is -normal/dr
-        vecs_s = [coeff_s * normal, -coeff_s * outer_w0[i] * normal]
-        if out1 != out0 or outer_w1[i] != 0.0:
-            rows_s.append(int(out1))
-            vecs_s.append(-coeff_s * outer_w1[i] * normal)
-
-        shape_part = (np.asarray(rows_s, dtype=int), np.asarray(vecs_s, dtype=float))
-
-        # dC/dt_out component (tilt)
-        rows_tout = [int(row) for row in target_rows]
-        vecs_tout = [coeff_t * float(w) * r_dir for w in target_weights]
-        tout_part = (
-            np.asarray(rows_tout, dtype=int),
-            np.asarray(vecs_tout, dtype=float),
-        )
-
-        # Row 1: (Shape, None, Tilt_Out)
-        constraints.append((shape_part, None, tout_part))
-
-        # 2. Inner constraint: t_in . r_dir - (theta_B - phi) = 0
-        # C = t_in . r_dir + phi - theta_B = 0
-        # Partials: dC/dt_in = r_dir, dC/dx = +d(phi)/dx
-        if disk_rows is not None and disk_r_hat is not None:
-            # dC/dx = +d(phi)/dx => rim component is -normal/dr, out is +normal/dr
-            vecs_s_in = [-coeff_s * normal, coeff_s * outer_w0[i] * normal]
-            if out1 != out0 or outer_w1[i] != 0.0:
-                vecs_s_in.append(coeff_s * outer_w1[i] * normal)
-            shape_part_in = (
-                np.asarray(rows_s, dtype=int),
-                np.asarray(vecs_s_in, dtype=float),
-            )
-
-            if theta_scalar is not None and _use_disk_theta_targeting(
-                global_params, matching_mode
-            ):
-                disk_target = _disk_theta_rows_weights_and_direction(
-                    data=data,
-                    i=i,
-                    theta_scalar_active=True,
-                )
-                if disk_target is None:
-                    continue
-                rows_disk, weights_disk, dirs_disk = disk_target
-                rows_tin = list(rows_disk)
-                vecs_tin = [
-                    coeff_t * float(w) * np.asarray(vec, dtype=float)
-                    for w, vec in zip(weights_disk, dirs_disk)
-                ]
-            else:
-                rows_tin = [int(row) for row in target_rows]
-                vecs_tin = [coeff_t * float(w) * r_dir for w in target_weights]
-                disk_target = _disk_theta_rows_weights_and_direction(
-                    data=data,
-                    i=i,
-                    theta_scalar_active=False,
-                )
-                if disk_target is not None:
-                    rows_disk, _weights_disk, dirs_disk = disk_target
-                    rows_tin.extend(rows_disk)
-                    vecs_tin.extend(
-                        [-coeff_t * np.asarray(vec, dtype=float) for vec in dirs_disk]
-                    )
-
-            tin_part = (
-                np.asarray(rows_tin, dtype=int),
-                np.asarray(vecs_tin, dtype=float),
-            )
-            # Row 2: (Shape, Tilt_In, None)
-            constraints.append((shape_part_in, tin_part, None))
-
-    return constraints or None
-
-
-def constraint_gradients_array(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: dict[int, int],
-) -> list[np.ndarray] | None:
-    """Return dense constraint gradients for rim matching (shape only)."""
-    row_constraints = constraint_gradients_rows_array(
-        mesh,
-        global_params,
-        positions=positions,
-        index_map=index_map,
-    )
-    if not row_constraints:
-        return None
-    constraints: list[np.ndarray] = []
-    for rows, vecs in row_constraints:
-        g_arr = np.zeros_like(positions)
-        np.add.at(g_arr, rows, vecs)
-        constraints.append(g_arr)
-    return constraints or None
-
-
-def constraint_gradients_rows_array(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: dict[int, int],
-) -> list[tuple[np.ndarray, np.ndarray]] | None:
-    """Return sparse-row constraint gradients for rim matching (shape only)."""
-    _ = index_map
-    data = _build_matching_data(mesh, global_params, positions)
-    if data is None:
-        return None
-
-    rim_rows = data["rim_rows"]
-    outer_rows = data["outer_rows"]
-    outer_idx0 = data["outer_idx0"]
-    outer_idx1 = data["outer_idx1"]
-    outer_w0 = data["outer_w0"]
-    outer_w1 = data["outer_w1"]
-    disk_rows = data["disk_rows"]
-    weight_sqrt = data["weight_sqrt"]
-    inv_dr = data["inv_dr"]
-    valid = data["valid"]
-    normal = data["normal"]
-    matching_mode = _resolve_matching_mode(global_params)
-
-    constraints: list[tuple[np.ndarray, np.ndarray]] = []
-    # Note: theta_scalar is available but does not change the gradient structure
-    # for the asymmetric constraints:
-    # 1. t_out . r = phi  => C = t_out.r - phi = 0
-    #    Gradient is -grad(phi).
-    # 2. t_in . r = theta_B - phi => C = t_in.r + phi - theta_B = 0
-    #    Gradient is +grad(phi).
-
-    rows_out_all: list[int] = []
-    vecs_out_all: list[np.ndarray] = []
-    rows_in_all: list[int] = []
-    vecs_in_all: list[np.ndarray] = []
-
-    for i, ok in enumerate(valid):
-        if not ok or weight_sqrt[i] == 0.0:
-            continue
-        coeff = weight_sqrt[i] * inv_dr[i]
-
-        out0 = outer_rows[outer_idx0[i]]
-        out1 = outer_rows[outer_idx1[i]]
-        rows_out = [int(rim_rows[i]), int(out0)]
-        vecs_out = [coeff * normal, -coeff * outer_w0[i] * normal]
-        if out1 != out0 or outer_w1[i] != 0.0:
-            rows_out.append(int(out1))
-            vecs_out.append(-coeff * outer_w1[i] * normal)
-
-        if matching_mode == "ring_average_radial_v1":
-            rows_out_all.extend(rows_out)
-            vecs_out_all.extend(vecs_out)
-        else:
-            constraints.append(
-                (
-                    np.asarray(rows_out, dtype=int),
-                    np.asarray(vecs_out, dtype=float),
-                )
-            )
-
-        if disk_rows is not None:
-            rows_in = [int(rim_rows[i]), int(out0)]
-            vecs_in = [-coeff * normal, coeff * outer_w0[i] * normal]
-            if out1 != out0 or outer_w1[i] != 0.0:
-                rows_in.append(int(out1))
-                vecs_in.append(coeff * outer_w1[i] * normal)
-            if matching_mode == "ring_average_radial_v1":
-                rows_in_all.extend(rows_in)
-                vecs_in_all.extend(vecs_in)
-            else:
-                constraints.append(
-                    (
-                        np.asarray(rows_in, dtype=int),
-                        np.asarray(vecs_in, dtype=float),
-                    )
-                )
-
-    if matching_mode == "ring_average_radial_v1":
-        if rows_out_all:
-            constraints.append(
-                (
-                    np.asarray(rows_out_all, dtype=int),
-                    np.asarray(vecs_out_all, dtype=float),
-                )
-            )
-        if rows_in_all:
-            constraints.append(
-                (
-                    np.asarray(rows_in_all, dtype=int),
-                    np.asarray(vecs_in_all, dtype=float),
-                )
-            )
-
-    return constraints or None
-
-
-def constraint_gradients_tilt_array(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: dict[int, int],
-    tilts_in: np.ndarray | None = None,
-    tilts_out: np.ndarray | None = None,
-) -> list[tuple[np.ndarray | None, np.ndarray | None]] | None:
-    """Return dense constraint gradients for rim matching (tilts)."""
-    row_constraints = constraint_gradients_tilt_rows_array(
-        mesh,
-        global_params,
-        positions=positions,
-        index_map=index_map,
-        tilts_in=tilts_in,
-        tilts_out=tilts_out,
-    )
-    if not row_constraints:
-        return None
-    constraints: list[tuple[np.ndarray | None, np.ndarray | None]] = []
-    for in_part, out_part in row_constraints:
-        g_in = None
-        g_out = None
-        if in_part is not None:
-            rows_in, vecs_in = in_part
-            g_in = np.zeros_like(positions)
-            np.add.at(g_in, rows_in, vecs_in)
-        if out_part is not None:
-            rows_out, vecs_out = out_part
-            g_out = np.zeros_like(positions)
-            np.add.at(g_out, rows_out, vecs_out)
-        constraints.append((g_in, g_out))
-    return constraints or None
-
-
-def constraint_gradients_tilt_rows_array(
-    mesh: Mesh,
-    global_params,
-    *,
-    positions: np.ndarray,
-    index_map: dict[int, int],
-    tilts_in: np.ndarray | None = None,
-    tilts_out: np.ndarray | None = None,
-) -> (
-    list[
-        tuple[
-            tuple[np.ndarray, np.ndarray] | None,
-            tuple[np.ndarray, np.ndarray] | None,
-        ]
-    ]
-    | None
-):
-    """Return sparse-row constraint gradients for rim matching (tilts)."""
-    _ = index_map, tilts_in, tilts_out
-    data = _build_matching_data(mesh, global_params, positions)
-    if data is None:
-        return None
-
-    disk_rows = data["disk_rows"]
-    disk_r_hat = data["disk_r_hat"]
-    weight_sqrt = data["weight_sqrt"]
-    valid = data["valid"]
-    theta_scalar = data["theta_scalar"]
-    matching_mode = _resolve_matching_mode(global_params)
-
-    normals = mesh.vertex_normals(positions=positions)
-
-    constraints: list[
-        tuple[
-            tuple[np.ndarray, np.ndarray] | None,
-            tuple[np.ndarray, np.ndarray] | None,
-        ]
-    ] = []
-
-    agg_out_rows: list[int] = []
-    agg_out_vecs: list[np.ndarray] = []
-    agg_in_rows: list[int] = []
-    agg_in_vecs: list[np.ndarray] = []
-
-    for i, ok in enumerate(valid):
-        if not ok or weight_sqrt[i] == 0.0:
-            continue
-        coeff = weight_sqrt[i]
-
-        target = _tilt_target_rows_weights_and_direction(
-            data=data,
-            positions=positions,
-            normals=normals,
-            i=i,
-            matching_mode=matching_mode,
-        )
-        if target is None:
-            continue
-        target_rows, target_weights, r_dir = target
-        out_part = (
-            np.asarray(target_rows, dtype=int),
-            np.asarray([coeff * float(w) * r_dir for w in target_weights], dtype=float),
-        )
-
-        if disk_rows is None or disk_r_hat is None:
-            if matching_mode == "ring_average_radial_v1":
-                agg_out_rows.extend(out_part[0].tolist())
-                agg_out_vecs.extend(out_part[1])
-            else:
-                constraints.append((None, out_part))
-            continue
-
-        if theta_scalar is not None and _use_disk_theta_targeting(
-            global_params, matching_mode
-        ):
-            disk_target = _disk_theta_rows_weights_and_direction(
-                data=data,
-                i=i,
-                theta_scalar_active=True,
-            )
-            if disk_target is None:
-                if matching_mode == "ring_average_radial_v1":
-                    agg_out_rows.extend(out_part[0].tolist())
-                    agg_out_vecs.extend(out_part[1])
-                else:
-                    constraints.append((None, out_part))
-                continue
-            rows_disk, weights_disk, dirs_disk = disk_target
-            rows_in = list(rows_disk)
-            vecs_in = [
-                coeff * float(w) * np.asarray(vec, dtype=float)
-                for w, vec in zip(weights_disk, dirs_disk)
-            ]
-        else:
-            rows_in = [int(row) for row in target_rows]
-            vecs_in = [coeff * float(w) * r_dir for w in target_weights]
-            disk_target = _disk_theta_rows_weights_and_direction(
-                data=data,
-                i=i,
-                theta_scalar_active=False,
-            )
-            if disk_target is not None:
-                rows_disk, _weights_disk, dirs_disk = disk_target
-                rows_in.extend(rows_disk)
-                vecs_in.extend(
-                    [-coeff * np.asarray(vec, dtype=float) for vec in dirs_disk]
-                )
-
-        in_part = (
-            np.asarray(rows_in, dtype=int),
-            np.asarray(vecs_in, dtype=float),
-        )
-        if matching_mode == "ring_average_radial_v1":
-            agg_out_rows.extend(out_part[0].tolist())
-            agg_out_vecs.extend(out_part[1])
-            agg_in_rows.extend(in_part[0].tolist())
-            agg_in_vecs.extend(in_part[1])
-        else:
-            constraints.append((None, out_part))
-            constraints.append((in_part, None))
-
-    if matching_mode == "ring_average_radial_v1":
-        if agg_out_rows:
-            constraints.append(
-                (
-                    None,
-                    (
-                        np.asarray(agg_out_rows, dtype=int),
-                        np.asarray(agg_out_vecs, dtype=float),
-                    ),
-                )
-            )
-        if agg_in_rows:
-            constraints.append(
-                (
-                    (
-                        np.asarray(agg_in_rows, dtype=int),
-                        np.asarray(agg_in_vecs, dtype=float),
-                    ),
-                    None,
-                )
-            )
-
-    return constraints or None
 
 
 def enforce_tilt_constraint(mesh: Mesh, global_params=None, **_kwargs) -> None:
@@ -1841,13 +281,7 @@ def enforce_tilt_constraint(mesh: Mesh, global_params=None, **_kwargs) -> None:
 
 
 def enforce_constraint(mesh: Mesh, global_params=None, **_kwargs) -> None:
-    """Project interface-shell heights onto the current rim law.
-
-    This is a hard shape realization of the existing interface law. It adjusts
-    the explicit outer secant shell along the interface normal so that the
-    discrete secant slope `phi` matches the current coupled tilt state as
-    closely as possible before the tilt-only projection runs.
-    """
+    """Project interface-shell heights onto the current rim law."""
     matching_mode = _resolve_matching_mode(global_params)
     context = str(_kwargs.get("context") or "").strip().lower()
     if matching_mode not in {
@@ -2050,6 +484,7 @@ def enforce_constraint(mesh: Mesh, global_params=None, **_kwargs) -> None:
 __all__ = [
     "coarse_rim_family_diagnostics",
     "constraint_gradients_array",
+    "constraint_gradients_joint_array",
     "constraint_gradients_rows_array",
     "constraint_gradients_tilt_array",
     "constraint_gradients_tilt_rows_array",
